Provide MWE for teststream.c replacement
Previously, the compiled binary of teststream was out-of-tree, with
references scattered throughout the Makefile and Python files to its
existence and use. Due to work on the eventual provision of in-flight
decimation, Python is no longer suitable as the core language in this
repository, as ioctl is required. Rust has been chosen to fulfil this
need.

This commit thus provides a minimal working rust example that replicates
the most basic functionality in teststream, with the necessary
dependencies and core logic demonstrated, especially in the context of
ioctls from the nix crate.

This provides the initial Cargo workspace setup.

Co-authored-by: ArtemiszenN <artemiszenn@gmail.com>