Add PythonMath trait
Due to the Python runner.py using eval to handle math expressions in the
ini file. we add a trait called PythonMath to handle such a similar
case. However, the eval function greedily coerces into floats as
specified by the documentation, and so, using our strong assumption of
base 2 exponentiation, we instead perform left shift with 1.