Add ConfigParser get values after ExtendedInterpolation
Python's ConfigParser allows for reading of values with
"ExtendedInterpolation", where you can replace in ${...} with values
that have the value. This is not part of the configparser create yet. We
add this as it is used in the Paths section of the configuration ini
file.

Currently, the error of values not being present from section/keys being
referenced but is absent is not being propagated out of a closure, and
will have to be fixed subsequently.