Migrate USB Engine details into library with MWE rewrite
We place the USB engine start/stop etc methods as being impls associated
to the VHF struct. One must take special care to ensure that the
variables being dropped at the end of scope does not result in a
double-free of the existing FD to the USB engine.

Corresponding Error types are also introduced.