Add rudimentary mmap fetch streaming
This example code is able to be executed once with data parity to the C
implementation. However, second run with this rust executable leads to
the wrong result, in which
a) Cleaning up with clear-FIFO fixes the issue or;
b) Running the C implementation once also fixes the issue.
As such, it is possible that the state teardown during lifetime scope
closure is likely to be improper for reasons yet to be known.
Nonetheless, this commit shows that there is almost identical core logic
being translated to Rust from C.

Basic features such as dynamic headers by runtime has not been
implemented.