Add native endian u64 to and from IQM Triplet
While data is being processed, it is done in the native endianness of
the machine. However, we must just keep in mind that the data that is
being read from the MMap and the files are of little endian.

This commit provides the From traits for both &u64, u64, IQMTriplet and
&IQMTriplet. Operating an AoS or SoA of u64/IQMTriplets will be
introduced into the relevant namespace in the future as necessary.