Define Page type for VHF Iterator Item
We introduce two different type of empty pages, the first serving
purpose similar to numpy's convolve with mode = 'full' for the start of
stream, and the second denoting the end of stream. This is necessary due
to how the start of stream is a special case compared to the "steady
state", where we require the function transforming a window of the
stream having to distinguish between the start and middle of the stream.

Currently, this may result in memory fetch fragmentation between pages,
and cache-line eviction during AtomicUsize synchronization between Arcs
across different threads, but we will have to profile this impact in the
future.