Warn instead of terminate for failed length assertion
Despite commit 37995bc41, some files still raise the assertion error
that the timings class duration length is not consistent with the
expected number of bytes as by file length. For now, we just catch and
log the error, as all of them are still identically just off by 1. The
files which fail these have to be investigated separately in the future.