Add sparse_m_delta CSV summary of all non-tmp/test files
This is just a fairly quick and dirty script that fetch all files within
the NAS that ends with bin without test or tmp anywhere in the name,
before trying to get the corresponding sparse_m_delta list.

There are 5844 files at the moment, which takes about 20 hours to run
across all of them.

The only gotchas so far is the fact that globbing for files vs fd seems
to roughly take the same amount of time, and the worker needs to reset
the parser variable unless the worker is destroyed.