Add VHF struct spawning of MMapReader thread
MmapReader has to consum the Mmap, so self.readback is instead passed as
ownership to the thread (which has JoinHandle<_>) type.
VHF struct therefore also now has to have the Arc<Mutex<_>> necessary to
receive the pages from the thread.

In the meantime, we are hardcoding the number of pages to read instead
of reading from configuration.