Add Condvar signaller between MmapReader and VHF
This conditional variable is meant to be used to unblock the
Iterator.next() method whilst waiting for more pages to come in.

This Condvar could probably be placed into the same Arc as the
Mutex<Pages>, but we'll determine its impact on having lifetime generics
in the struct definition when the time for it comes. (See condvar docs
using &*Arc<(Mutex<_>, Condvar) to synchronise the two threads.)