Implement Iterator trait on VHF struct
With all the primitives finally laid out, we can properly define the
next method, which yields an "index" and a "window" of data for any
transforms to work window-wise. As hinted, the current window size is
still arbitrary, and needs to be further tested.

There still currently is a concern that the waiting for buffer to
populate before emitting out in `.next()` might be blocking on methods
further down the loop, for things such as writing to file etc.

The upper bound as provided in the size hint is still only an estimate
and not fully proven.