Terminate iterator fetching from buffer early if stalled
Whilst writing the iterator unit test, an improper configuration of
number pages and windows led to the unit tests not completing. In
reality, this might translate to a case where the mmap_thread has
panicked. In the future, VHF as a struct has to check that the
corresponding JoinHandle if it has panicked or not, and to set the
engine_running to false if found to have panicked instead.