Migrate process constants to board_ioctl_consts
There are various constants under the process submodule, namely,
constants related to board interaction, and constants related to
implementation details shared across submodules of process submodule.

This commit therefore vacates the consts namespace for future commit for
the purposes of sharing of "more generic" constants rather than hunting
down for the location it was initially shared into with.