Add WriteBlock type
The goal of WriteBlock as a struct is for there to be a common output
from any ChunkedStream maps that is passed to what is eventually
different writer implementations.

The struct will allow for optionalities such as rollover index relative
to the start of the WriteBlock, the sign of the rollovers etc, as needed
in the future.