Add non-overlapping identity transform for stream
We define a struct which full details all necessary information about
how to appropriately step_by the `.iter()` of VHF, and the corresponding
functions to apply on each window before passing to either a file
collector or a reducer.
A reducer can be used in the case for something such as aggregating an
average of a FFT for example, and the file collector as a separate
example would take a transformed stream for writing.
In the case where no reduction/transformation (i.e.: cases of fold) is
desired, we utilise this non-overlapping transform in place of what
should be interpreted as noop in the context of the stream.