Fix wrong termination of engine_running
fetch_not places the result back into the AtomicBool, which is not what
we want.

This commit causes a regression in unit-tests, but is necessary in the
case of a live fetching from the FPGA. Eventually. all the unit tests
should have the debug_vhf in their own thread slowly having pushing
pages into buffer instead before configuring the engine_running to be
switched off.