Change MWE stream towards real usage
The code has shaped up enough that the stream file now looks closer to
what was intended on the binary end.

However, VHF currently still does not implement iter() method, so the
lifetime is consumed. The termination of VHF is called from within
next() for now, but should be delegated separately.