Add padding to end of stream
The idea of this is for the the final window of data to be at least be
emitted to pass through the `step_by(..)` to pass to the writer.
However, it might not be entirely correct so far, as the last file still
appears to be having less than expected number of elements.