Change config reading functions to be more functional
Matching previously was quite cumbersome.
Also actually used defaults in places that were logged rather just
logging out that the default was being used.
Finally, also followed the principle of unrepresentable state as much as
possible, and canonicalized the path before keeping into the struct.
This canonicalized path is then saved in the VHF file's metadata instead
of the symblink too.