Change unit tests to use pages from iterator function
This will allow us to gradually place pages into the buffer rather than
having to collect everything into a vector and one shot dump everything
in.
Furthermore, this creates a much more consistent way of specifying the
delay between pages placed onto the buffer.
Take special care to account for the fact that the original child thread
is also trying to acquire the same Mutex.