Fix not being able to write more than 2 GiB
This was initially mistaken as having a deadlock in the writer around
the second file. This fix instead makes it clear that wrapping_sub wraps
around i32::MIN/MAX when the intention is in fact to wrap up and across
i32::MAX to 0.

In addition, also checks that `next` has not already overtaken `"prev"`
in the circular buffer.