Push file-writing to a separate thread in stream
This really ties to make sure that the back log of file writing for slow
NAS is never the culprit for draining the fixed sized SyncChannel (#31)
between the parent VHF thread and child MMapReader thread.

It is not clear if this commit will still keep the intended behaviour of
panicking when there's too much data being dumped out the VHF booard
because it has already been having too much data on the FPGA buffer.