Add `fuser` check for VHF board
This is meant for resolving to the interface path to see if the board is
being used by any process.
Correspondingly, expose `Board` struct outside of `clear-fifo` feature
since `fuser` check will be useful. That said, `clear-fifo` feature for
exposing number of elements flushed from FPGA FIFO in log is still
mentioned.