Add mapping of `./stream` argument flags to Configs struct
This will handle the ArgMatches output into Configs struct.

We try to defer as much validation logic as possible to other functions,
but we try to not place "invalid" state into the struct, i.e.: best
effort to have paths be canonical before exiting the scope responsible
for mutating that field.