Invert StreamFold and StreamFoldParameters relation
Previously, StreamFold was a enum container what was effectively the
same StreamFoldParameters object. It just is more natural to have the
parameters include the fold type in a field.

This commit thus places StreamFold as field in StreamFoldParameters, and
then renames StreamFold to StreamFoldOp, followed by
StreamFoldParameters renamed to StreamFold.