Refactor Config to have BoardConfig
Config as a struct can no longer guarantee that it contains valid state
representation of all relevant properties of Board / Processing/
File-Writing. As such, it is most sensible for Config to remain as the
"staging" area for processing VHF_board_param.ini and CLI argument flags
in a best effort, before creating (possibly failing) a board_config
(this commit) and file-writer builder/config in the future.