Change v1_writer to accept its own specific arguments
Each of the various writers need a different set of arguments, and so
having a specific FileConfig for all `impl VHFWriter` is probably not
the wisest. This commit aims to first move just V1 file writer, while
leaving v1 stdout and stream changes separately.

This breaks the building of stream and is therefore in its own branch.
Please see how to bisect across a branch using --first-parent.