Change identity_overflow to encode by absolute position in VHFIterStream
Previously, indices of overflow was relative to the VHFIter Window,
which meant that the writer had to instead keep track on the number of
blocks seen (highly complicated). As such, we differ to the StreamOp to
take up the responsibility of ensuring that elements from the
StreamOp::func instead had their indices almost as if it was enumerated
straight from the FPGA a RawVHFWord at a time.
As noted in a comment, u64 elements even at 80MHz will take many many
years to hit, so its entirely safe.

Also, update stepped_overlapping_identity_b to also use a larger stream.