Add v2_writer writing M_OVERFLOW_IDX as index relative to file start
After having changed the M_OVERFLOW_RAW to be indexed relative to start
of VHFIter, it now changes the offset handling M_OVERFLOW_RAW by number
of blocks of M_OVERFLOW_RAW seen to just having to offset to relative to
file.
Slight concern with regards to potential atomic fetch.