Fix Polar to VHFWord
Thinking about it harder, it is unlikely that it was the Python parsing
that parsed the Raw word wrongly, but it was the generation of Words in
Rust that was done wrongly. Surely enough, the unit testing was
insufficient, and a lot of edge cases were caught.

It is important to think of the partial open intervals of all
operations.

At present, there still is some issue for phases where there is m values
outside of i16 interval, but more will be done.