Fix Python m_overflow being off by 1 for first update
Turns out, the rust unit test was so small that it triggered the else
branch that I guess was not one to occur often enough. By setting the
default block_size to a smaller value, the off-by-1 in idx, delta was
found, and so this fix was made. The Rust unit tests now pass without
issue.