Add initial v2 ManifoldManager
Needs more unit-testing still. This will read from the relevant declared
region of the file and store into its internal form similar to how it
was done in v1.
Currently, try_cast_slice will heap allocate the entire file all at once
to get the &[u64], which is not good. This can be tested by changing the
test/v2_write.rs file to create a larger file.
Also, a subtle bug was caught where if the declared region for
m-overflows was 0 in length, which results in try_cast_slice erroring
out, and is avoided in this commit.