Add parser test for v2 writer
This also checks the reduced phase as being read out from the parser.
More tests will be made for handling the case of m-overflow.
More tests will be needed for handling software filters.

mem::drop, or scoping is necessary, as the parser will read from the
file, before the file writer flushes out it's m-overflow idx.