Add representations of window functions
Looked into Scipy documentation for a reasonably exhaustive set of
representation of windows that can be used. This will be the numerical
representation of the like of window functions that are then used in
filter methods. There will be no attempt in the future to ensure that
the values passed into here (and related structs) are accurate. The only
goal of them is for serialization for recording purposes.