Fix v2 file writing tests to use global VHF_MMAP_WINDOW_LEN constant
Whilst considering what it'd take to have a decimation factor of >500,
it became apparent that the existing window length meant that only 20 or
so data points would have been emitted at each iteration step... which
was unwieldy.
The cyclic requirement of the tests were then found to have these hard
coded values of 19 to begin with, which would break raising the constant
of 20.