Add doc-images into fold documentation
Placed behind a feature flag as a Make step is necessary for creating
the PNGs that from tex before the "embed-doc-image" dependency "links"
all the necessary images from `target/docs/` to the appropriate
locations where the images are generated for the images to show up in
the browser where all the documentation resides.

This is just scoping the macro in as specified in the embed-doc-image
dependency to ensure that it works.