Remove wrong assertion in span_as_offset
It is entirely possible that the duration argument passed from the user
is not a nice multiple of the sample interval. We follow the strategy
that was used in python to just round blindly.

We follow the wording in the documentation and clamp the value emitted
as an index accordingly to between 0 and the trace_len.