Check serialization of user-omitted flags
This is particular so for *_enabled flags in the configuration file
being set to false should result in a `null` serialization, which
denotes that user has deliberately chosen to omit the specified field.

However, the VHF board will still write related `-g` and `f` fields at
FPGA initialization time with 0s to ensure that no stale configuration
is left behind.