Add packing of filtered phase into write block
Adds the function and control flow around the core logic of taking what
is expected to be filtered unwrapped phase to be passed to the writer.
As such, an `idx` to specify the 0th element received by this
`pack_into_write_block` should correspond to which element in the
stream.