Add logic to filtfilt closure in `func` field of StreamFold
This is the broad strokes of what has to be done, namely, unwrapping,
filtering, followed by packing back into VHFWord for writing.
Parameters associated to step sizes are not yet done, and left as
comments as "todos" to check for compilation.

While it is possible to do a trivial .skip().step_by() on the radius, it
is decided that "averaging" according to the weights of filter weights
`b` and `a` is preferable, which is no different from just running the
`filtfilt` on the radius instead.
This is consistent with the initial intuition that filtering solely on
just I and Q parameters should have been a valid process. (Note that
this naive process in invalid because the value of m is often changing
too much in all data we collect.)