Add diagram for rejecting data from initial padding
This diagram serves to demonstrate when data filtered from the input
should not be written into the output. Simply put, if any input array
element is passed through the filter, then it must be written out.

The z-order diagram is inverted to ensure is transparency is respected
for the final output.

The makefile is modified to not stall in interactive mode on errors.