Update `num_before_first_drop` in `MapArg` to use usize
It was an oversight to expect that there will always be a non-zero
number of elements that are dropped prior to file writing.
Further clarification on how these MapArg fields are meant to be used.

Note that `num_before_first_drop` is typically a function of the other
field `effective_decimation` and out-of-struct parameter - `filter_len`.

The consequential `words_dropped_before_first_write` is thus also
updated to remove unnecessary `.(try_)into()` calls.