Fix unwrap_phases_in_windows
This was initially misunderstood as the wrong overflow_idx being
generated by `pack_into_write_block`, but after checking that the output
and input of `filtfilt`, it became apparent the problem was coming from
before, not after.
Currently, it still stands that the indices generated by the closure is
nonsensical.