Improve Overflow tuple in `dbg!` macro
`dbg!` macro uses `{:#?}` under the hood, which means that trying to
determine the issue is abit harder as the output is much longer than it
needs to be. Hijack this by using `write!` macro. (Not sure why the
`cfg(...)` appears to be inverted for it to use the `write!` in `cargo
test`.)