Add other filtfilt tests
These tests only vary the length of the FIR filter passed to filtfilt,
and has yet to test the decimation_factor parameter.
It is currently constructed in a way where it abuses the knowledge that
the signal have a wavelength of 2, and so FIR filters are effectively
invariant over the input when creating the output.

The tests currently are failing, more tweaking on `write_idx` and
`initial_skip` from within the closure is needed, or how the
`filtfilt_f64` function does it. It seems quite apparent that the first
sign arising from the element prior to the first element being written
is discarded wrongfully.

More work in the future will also has to be done if `filtfilt_f64` will
be paralliseable by not having a Mutex over the FFTPlanner; or if
Arc<dyn FnMut> implementing DerefMut etc in the scoped closure is
possible.